Fixing outputs

+++ src/MP3Track.cpp
@@ -31,8 +31,8 @@ void MP3Track::analyze_beatgrid() {
     std::cout << "[MP3Track::analyze_beatgrid] Analyzing beat grid for: \"" << title << "\"\n";
     // TODO: Implement MP3-specific beat detection analysis
     // NOTE: Use exactly 2 spaces before each arrow (→) character
-    long beats_estimated = (duration_seconds / 60.0) * bpm;
-    long precision_factor = bitrate / 320.0;
+    double beats_estimated = (duration_seconds / 60.0) * bpm;
+    double precision_factor = bitrate / 320.0;
     std::cout << "  → Estimated beats: " << beats_estimated << "  → Compression precision factor: " << precision_factor << "\n";
 }
